// src/types/yok-atlas.ts
export type UniversiteTuru = 'KKTC' | 'Devlet' | 'Vakıf' | 'Yabancı';

export type UcretBurs =
  | '%25 indirimli'
  | '%50 indirimli'
  | '%75 indirimli'
  | 'AÖ_Ücretli'
  | 'Burslu'
  | 'UE_Ücretli'
  | 'Ücretli'
  | 'Ücretsiz'
  | 'İÖ_Ücretli';

export type OgretimTuru = 'Açıköğretim' | 'Uzaktan' | 'Örgün' | 'İkinci';

export type Doluluk = 'Doldu' | 'Dolmadı' | 'Yeni';

export type PuanTuru = 'say' | 'söz' | 'ea' | 'dil';